Convert pycurl -> httpx
